latona: power: don't detect full charge with CHG_ING GPIO
CHG_ING is high when the battery is fully charged, but we
want to stop charging when the soc value is at least 95.
We hence don't use CHG_ING to determine whether the battery is
fully charged or not, but we check the voltage and the soc value.
Once the battery is considered fully charged, wait for at least
limit_recharging_time seconds before recharging.

Change-Id: I2807fc0efa323aeb3853a5646ebc4c17f4e8d842
Signed-off-by: David Rodero <david@codiguiel.net>