drivers: omap-serial: hold wakelock when serial uart driver resumes
Hold a timed wakelock upon resume that would be sufficient to allow
the userspace to handle the Serial UART RX data.

UART_RX signal is wakeup enabled, but when the wakeup occurs, the
system is put back to suspend immediately because of "unknown wakeup".
Due to this, the system doesn't repond properly to UART RX Data like
Serial Console input and Bluetooth RX (Pairing Request, File Transfer
Request and AVRCP) during suspend.